+++ src/wkg.webview.cpp
@@ -268,7 +268,7 @@ namespace saucer
         webkit_web_view_evaluate_javascript(platform->web_view, code.c_str(), -1, nullptr, nullptr, nullptr, nullptr, nullptr);
     }
 
-    std::uint64_t impl::inject(const script &script) // NOLINT(*-function-const)
+    std::size_t impl::inject(const script &script) // NOLINT(*-function-const)
     {
         using enum load_time;
         using enum web_frame;
@@ -305,7 +305,7 @@ namespace saucer
         }
     }
 
-    void impl::uninject(std::uint64_t id) // NOLINT(*-function-const)
+    void impl::uninject(std::size_t id) // NOLINT(*-function-const)
     {
         if (!platform->scripts.contains(id))
         {
